refactor: move labelNode func into util.go
